Backend/backend-ml/requirements.txt
fastapi==0.73.0
gunicorn==20.1.0
numpy
uvicorn==0.15.0
image
tensorflow-cpu==2.7.0
tensorflow-hub
wget
wheel
python-multipart
urllib3==1.26.6

pandas